Max patch for localisation testing
For general purpose, users may use this resource in any way they see fit. For the purpose of the dissertation assessment, this patch was created and used during a binaural localisation test. The interface (presentation view) contains a frame for MIRA, therefore in order to run this patch the way it was intended, one must have access to an external device (iPad for the project) running MIRA for Max MSP